requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.2
auto-py-to-exe==2.8.0
beautifulsoup4==4.9.3
binary==1.0.0
bootstrap4==0.1.0
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.6
django-bootstrap4==3.0.1
django-countries==7.2.1
django-decouple==2.1
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
docopt==0.6.2
docutils==0.17
Eel==0.12.4
et-xmlfile==1.1.0
filelock==3.0.12
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
gitignore==0.0.8
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
isort==5.7.0
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kivymd==0.104.1
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupPy==1.14
mccabe==0.6.1
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
pafy==0.5.5
pefile==2019.4.18
phonenumbers==8.12.20
Pillow==8.3.1
playsound==1.2.2
psycopg2==2.9.1
pycparser==2.20
Pygments==2.8.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pylint==2.6.0
pypiwin32==223
python-decouple==3.4
pytube==10.7.1
pytz==2021.1
pyuca==1.2
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
tk==0.1.0
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.3
whichcraft==0.6.1
whitenoise==5.3.0
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
youtube-dl==2021.3.14
zope.event==4.5.0
zope.interface==5.3.0
